revno: 5844 message:   I18n db of available/selected lists
committer: Lars Helge Overland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Wed 2012-02-01 11:50:31 +0100

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-organisationunit/src/main/java/org/hisp/dhis/oum/action/organisationunitgroup/GetOrganisationUnitGroupAction.java
@@ -91,7 +91,7 @@ public class GetOrganisationUnitGroupAction
     public String execute()
         throws Exception
     {
-        organisationUnitGroup = organisationUnitGroupService.getOrganisationUnitGroup( id.intValue() );
+        organisationUnitGroup = organisationUnitGroupService.getOrganisationUnitGroup( id );
 
         memberCount = organisationUnitGroup.getMembers().size();
