+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-organisationunit/src/main/java/org/hisp/dhis/oum/action/organisationunitgroup/RemoveOrganisationUnitGroupAction.java
@@ -68,7 +68,7 @@ public class RemoveOrganisationUnitGroupAction
         throws Exception
     {
         organisationUnitGroupService.deleteOrganisationUnitGroup( organisationUnitGroupService
-            .getOrganisationUnitGroup( id.intValue() ) );
+            .getOrganisationUnitGroup( id ) );
 
         return SUCCESS;
     }
