+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-organisationunit/src/main/java/org/hisp/dhis/oum/action/organisationunitgroupset/GetGroupSetAction.java
@@ -82,7 +82,7 @@ public class GetGroupSetAction
     public String execute()
         throws Exception
     {
-        organisationUnitGroupSet = organisationUnitGroupService.getOrganisationUnitGroupSet( id.intValue() );
+        organisationUnitGroupSet = organisationUnitGroupService.getOrganisationUnitGroupSet( id );
 
         if ( organisationUnitGroupSet != null )
         {
