+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-organisationunit/src/main/java/org/hisp/dhis/oum/action/organisationunitgroupset/RemoveGroupSetAction.java
@@ -68,7 +68,7 @@ public class RemoveGroupSetAction
         throws Exception
     {
         organisationUnitGroupService.deleteOrganisationUnitGroupSet( organisationUnitGroupService
-            .getOrganisationUnitGroupSet( id.intValue() ) );
+            .getOrganisationUnitGroupSet( id ) );
 
         return SUCCESS;
     }
